Move tests for Issue::create() in CreateTest
