+++ tests/Unit/Api/IssueTest.php
@@ -288,354 +288,6 @@ class IssueTest extends TestCase
     }
 
     /**
-     * Test cleanParams().
-     *
-     * @covers ::create
-     * @covers ::cleanParams
-     * @covers ::getIssueCategoryApi
-     * @covers ::getIssueStatusApi
-     * @covers ::getProjectApi
-     * @covers ::getTrackerApi
-     * @covers ::getUserApi
-     * @test
-     */
-    public function testCreateWithClientCleansParameters()
-    {
-        // Test values
-        $response = 'API Response';
-        $parameters = [
-            'project' => 'Project Name',
-            'category' => 'Category Name',
-            'status' => 'Status Name',
-            'tracker' => 'Tracker Name',
-            'assigned_to' => 'Assigned to User Name',
-            'author' => 'Author Name',
-        ];
-
-        // Create the used mock objects
-        $getIdByNameApi = $this->createMock('Redmine\Api\Project');
-        $getIdByNameApi->expects($this->exactly(3))
-            ->method('getIdByName')
-            ->willReturn('cleanedValue');
-        $issueCategoryGetIdByNameApi = $this->createMock('Redmine\Api\IssueCategory');
-        $issueCategoryGetIdByNameApi->expects($this->exactly(1))
-            ->method('getIdByName')
-            ->willReturn('cleanedValue');
-        $getIdByUsernameApi = $this->createMock('Redmine\Api\User');
-        $getIdByUsernameApi->expects($this->exactly(2))
-            ->method('getIdByUsername')
-            ->willReturn('cleanedValue');
-
-        $client = $this->createMock(Client::class);
-        $client->expects($this->exactly(5))
-            ->method('getApi')
-            ->willReturnMap(
-                [
-                    ['project', $getIdByNameApi],
-                    ['issue_category', $issueCategoryGetIdByNameApi],
-                    ['issue_status', $getIdByNameApi],
-                    ['tracker', $getIdByNameApi],
-                    ['user', $getIdByUsernameApi],
-                ]
-            );
-
-        $client->expects($this->once())
-            ->method('requestPost')
-            ->with(
-                '/issues.xml',
-                $this->logicalAnd(
-                    $this->stringStartsWith('<?xml version="1.0"?>' . "\n" . '<issue>'),
-                    $this->stringEndsWith('</issue>' . "\n"),
-                    $this->stringContains('<project_id>cleanedValue</project_id>'),
-                    $this->stringContains('<category_id>cleanedValue</category_id>'),
-                    $this->stringContains('<status_id>cleanedValue</status_id>'),
-                    $this->stringContains('<tracker_id>cleanedValue</tracker_id>'),
-                    $this->stringContains('<assigned_to_id>cleanedValue</assigned_to_id>'),
-                    $this->stringContains('<author_id>cleanedValue</author_id>')
-                )
-            )
-            ->willReturn(true);
-        $client->expects($this->exactly(1))
-            ->method('getLastResponseBody')
-            ->willReturn($response);
-
-        // Create the object under test
-        $api = new Issue($client);
-
-        // Perform the tests
-        $this->assertSame($response, $api->create($parameters));
-    }
-
-    /**
-     * @covers ::create
-     * @covers ::cleanParams
-     * @covers ::getIssueStatusApi
-     * @test
-     */
-    public function testCreateWithHttpClientRetrievesIssueStatusId()
-    {
-        $client = AssertingHttpClient::create(
-            $this,
-            [
-                'GET',
-                '/issue_statuses.json',
-                'application/json',
-                '',
-                200,
-                'application/json',
-                '{"issue_statuses":[{"name":"Status Name","id":123}]}'
-            ],
-            [
-                'POST',
-                '/issues.xml',
-                'application/xml',
-                '<?xml version="1.0"?><issue><status_id>123</status_id></issue>',
-                200,
-                'application/xml',
-                '<?xml version="1.0"?><issue></issue>'
-            ]
-        );
-
-        // Create the object under test
-        $api = new Issue($client);
-
-        $xmlElement = $api->create(['status' => 'Status Name']);
-
-        // Perform the tests
-        $this->assertInstanceOf(SimpleXMLElement::class, $xmlElement);
-        $this->assertXmlStringEqualsXmlString(
-            '<?xml version="1.0"?><issue></issue>',
-            $xmlElement->asXml(),
-        );
-    }
-
-    /**
-     * @covers ::create
-     * @covers ::cleanParams
-     * @covers ::getProjectApi
-     * @test
-     */
-    public function testCreateWithHttpClientRetrievesProjectId()
-    {
-        $client = AssertingHttpClient::create(
-            $this,
-            [
-                'GET',
-                '/projects.json',
-                'application/json',
-                '',
-                200,
-                'application/json',
-                '{"projects":[{"name":"Project Name","id":3}]}'
-            ],
-            [
-                'POST',
-                '/issues.xml',
-                'application/xml',
-                '<?xml version="1.0"?><issue><project_id>3</project_id></issue>',
-                200,
-                'application/xml',
-                '<?xml version="1.0"?><issue></issue>'
-            ]
-        );
-
-        // Create the object under test
-        $api = new Issue($client);
-
-        $xmlElement = $api->create(['project' => 'Project Name']);
-
-        // Perform the tests
-        $this->assertInstanceOf(SimpleXMLElement::class, $xmlElement);
-        $this->assertXmlStringEqualsXmlString(
-            '<?xml version="1.0"?><issue></issue>',
-            $xmlElement->asXml(),
-        );
-    }
-
-    /**
-     * @covers ::create
-     * @covers ::cleanParams
-     * @covers ::getIssueCategoryApi
-     * @test
-     */
-    public function testCreateWithHttpClientRetrievesIssueCategoryId()
-    {
-        $client = AssertingHttpClient::create(
-            $this,
-            [
-                'GET',
-                '/projects/3/issue_categories.json',
-                'application/json',
-                '',
-                200,
-                'application/json',
-                '{"issue_categories":[{"name":"Category Name","id":45}]}'
-            ],
-            [
-                'POST',
-                '/issues.xml',
-                'application/xml',
-                '<?xml version="1.0"?><issue><project_id>3</project_id><category_id>45</category_id></issue>',
-                200,
-                'application/xml',
-                '<?xml version="1.0"?><issue></issue>'
-            ]
-        );
-
-        // Create the object under test
-        $api = new Issue($client);
-
-        $xmlElement = $api->create(['project_id' => 3, 'category' => 'Category Name']);
-
-        // Perform the tests
-        $this->assertInstanceOf(SimpleXMLElement::class, $xmlElement);
-        $this->assertXmlStringEqualsXmlString(
-            '<?xml version="1.0"?><issue></issue>',
-            $xmlElement->asXml(),
-        );
-    }
-
-    /**
-     * @covers ::create
-     * @covers ::cleanParams
-     * @covers ::getTrackerApi
-     * @test
-     */
-    public function testCreateWithHttpClientRetrievesTrackerId()
-    {
-        $client = AssertingHttpClient::create(
-            $this,
-            [
-                'GET',
-                '/trackers.json',
-                'application/json',
-                '',
-                200,
-                'application/json',
-                '{"trackers":[{"name":"Tracker Name","id":9}]}'
-            ],
-            [
-                'POST',
-                '/issues.xml',
-                'application/xml',
-                '<?xml version="1.0"?><issue><tracker_id>9</tracker_id></issue>',
-                200,
-                'application/xml',
-                '<?xml version="1.0"?><issue></issue>'
-            ]
-        );
-
-        // Create the object under test
-        $api = new Issue($client);
-
-        $xmlElement = $api->create(['tracker' => 'Tracker Name']);
-
-        // Perform the tests
-        $this->assertInstanceOf(SimpleXMLElement::class, $xmlElement);
-        $this->assertXmlStringEqualsXmlString(
-            '<?xml version="1.0"?><issue></issue>',
-            $xmlElement->asXml(),
-        );
-    }
-
-    /**
-     * @covers ::create
-     * @covers ::cleanParams
-     * @covers ::getUserApi
-     * @test
-     */
-    public function testCreateWithHttpClientRetrievesUserId()
-    {
-        $client = AssertingHttpClient::create(
-            $this,
-            [
-                'GET',
-                '/users.json',
-                'application/json',
-                '',
-                200,
-                'application/json',
-                '{"users":[{"login":"Author Name","id":5},{"login":"Assigned to User Name","id":6}]}'
-            ],
-            [
-                'POST',
-                '/issues.xml',
-                'application/xml',
-                '<?xml version="1.0"?><issue><assigned_to_id>6</assigned_to_id><author_id>5</author_id></issue>',
-                200,
-                'application/xml',
-                '<?xml version="1.0"?><issue></issue>'
-            ]
-        );
-
-        // Create the object under test
-        $api = new Issue($client);
-
-        $xmlElement = $api->create(['assigned_to' => 'Assigned to User Name', 'author' => 'Author Name']);
-
-        // Perform the tests
-        $this->assertInstanceOf(SimpleXMLElement::class, $xmlElement);
-        $this->assertXmlStringEqualsXmlString(
-            '<?xml version="1.0"?><issue></issue>',
-            $xmlElement->asXml(),
-        );
-    }
-
-    /**
-     * Test create() and buildXML().
-     *
-     * @covers ::create
-     * @covers ::attachCustomFieldXML
-     * @test
-     */
-    public function testCreateBuildsXmlForCustomFields()
-    {
-        // Test values
-        $response = 'API Response';
-        $parameters = [
-            'custom_fields' => [
-                [
-                    'id' => 123,
-                    'name' => 'cf_name',
-                    'field_format' => 'string',
-                    'value' => [1, 2, 3],
-                ],
-            ],
-        ];
-
-        // Create the used mock objects
-        $client = $this->createMock(Client::class);
-
-        $client->expects($this->once())
-            ->method('requestPost')
-            ->with(
-                '/issues.xml',
-                $this->logicalAnd(
-                    $this->stringStartsWith('<?xml version="1.0"?>'),
-                    $this->stringContains('<issue>'),
-                    $this->stringContains('<custom_fields type="array">'),
-                    $this->stringContains('<custom_field name="cf_name" field_format="string" id="123" multiple="true">'),
-                    $this->stringContains('<value>1</value>'),
-                    $this->stringContains('<value>2</value>'),
-                    $this->stringContains('<value>3</value>'),
-                    $this->stringContains('</custom_field>'),
-                    $this->stringContains('</custom_fields>'),
-                    $this->stringEndsWith('</issue>' . "\n")
-                )
-            )
-            ->willReturn(true);
-        $client->expects($this->exactly(1))
-            ->method('getLastResponseBody')
-            ->willReturn($response);
-
-        // Create the object under test
-        $api = new Issue($client);
-
-        // Perform the tests
-        $this->assertSame($response, $api->create($parameters));
-    }
-
-    /**
      * Test update().
      *
      * @covers ::update
@@ -866,203 +518,6 @@ class IssueTest extends TestCase
 
         // Perform the tests
         $this->assertSame($response, $api->addNoteToIssue(5, 'Note content'));
-    }
-
-    /**
-     * Test buildXML().
-     *
-     * @test
-     */
-    public function testBuildXmlWithCustomFields()
-    {
-        // Test values
-        $parameters = [
-            'custom_fields' => [
-                ['id' => 225, 'value' => 'One Custom Field'],
-                ['id' => 25, 'value' => 'Second Custom Field'],
-                ['id' => 321, 'value' => 'http://test.local/?one=first&two=second'],
-            ],
-        ];
-
-        // Create the used mock objects
-        $client = $this->createMock(Client::class);
-        $client->expects($this->once())
-            ->method('requestPost')
-            ->with(
-                '/issues.xml',
-                $this->logicalAnd(
-                    $this->stringStartsWith('<?xml version="1.0"?>' . "\n" . '<issue>'),
-                    $this->stringEndsWith('</issue>' . "\n"),
-                    $this->stringContains('<custom_fields type="array">'),
-                    $this->stringContains('</custom_fields>'),
-                    $this->stringContains('<custom_field id="225"><value>One Custom Field</value></custom_field>'),
-                    $this->stringContains('<custom_field id="25"><value>Second Custom Field</value></custom_field>'),
-                    $this->stringContains('<custom_field id="321"><value>http://test.local/?one=first&amp;two=second</value></custom_field>')
-                )
-            );
-
-        // Create the object under test
-        $api = new Issue($client);
-
-        // Perform the tests
-        $api->create($parameters);
-    }
-
-    /**
-     * Test buildXML().
-     *
-     * @test
-     */
-    public function testBuildXmlWithWatchers()
-    {
-        // Test values
-        $parameters = [
-            'watcher_user_ids' => [5, 25],
-        ];
-
-        // Create the used mock objects
-        $client = $this->createMock(Client::class);
-        $client->expects($this->once())
-            ->method('requestPost')
-            ->with(
-                '/issues.xml',
-                $this->logicalAnd(
-                    $this->stringStartsWith('<?xml version="1.0"?>' . "\n" . '<issue>'),
-                    $this->stringEndsWith('</issue>' . "\n"),
-                    $this->stringContains('<watcher_user_ids type="array">'),
-                    $this->stringContains('</watcher_user_ids>'),
-                    $this->stringContains('<watcher_user_id>5</watcher_user_id>'),
-                    $this->stringContains('<watcher_user_id>25</watcher_user_id>')
-                )
-            );
-
-        // Create the object under test
-        $api = new Issue($client);
-
-        // Perform the tests
-        $api->create($parameters);
-    }
-
-    /**
-     * Test buildXML().
-     *
-     * @test
-     */
-    public function testBuildXmlWithUploads()
-    {
-        // Test values
-        $parameters = [
-            'uploads' => [
-                [
-                    'token' => 'first-token',
-                    'filename' => 'SomeRandomFile.txt',
-                    'description' => 'Simple description',
-                    'content_type' => 'text/plain',
-                ],
-                [
-                    'token' => 'second-token',
-                    'filename' => 'An-Other-File.css',
-                    'content_type' => 'text/css',
-                ],
-            ],
-        ];
-
-        // Create the used mock objects
-        $client = $this->createMock(Client::class);
-        $client->expects($this->once())
-            ->method('requestPost')
-            ->with(
-                '/issues.xml',
-                $this->logicalAnd(
-                    $this->stringStartsWith('<?xml version="1.0"?>' . "\n" . '<issue>'),
-                    $this->stringEndsWith('</issue>' . "\n"),
-                    $this->stringContains('<uploads type="array">'),
-                    $this->stringContains('</uploads>'),
-                    $this->stringContains(
-                        '<upload>'
-                        . '<token>first-token</token>'
-                        . '<filename>SomeRandomFile.txt</filename>'
-                        . '<description>Simple description</description>'
-                        . '<content_type>text/plain</content_type>'
-                        . '</upload>'
-                    ),
-                    $this->stringContains(
-                        '<upload>'
-                        . '<token>second-token</token>'
-                        . '<filename>An-Other-File.css</filename>'
-                        . '<content_type>text/css</content_type>'
-                        . '</upload>'
-                    )
-                )
-            );
-
-        // Create the object under test
-        $api = new Issue($client);
-
-        // Perform the tests
-        $api->create($parameters);
-    }
-
-    /**
-     * Test buildXML().
-     *
-     * @test
-     */
-    public function testBuildXmlWithWatcherAndUploadAndCustomFieldAndStandard()
-    {
-        // Test values
-        $parameters = [
-            'watcher_user_ids' => [5],
-            'subject' => 'Issue subject with some xml entities: & < > " \' ',
-            'description' => 'Description with some xml entities: & < > " \' ',
-            'uploads' => [
-                [
-                    'token' => 'first-token',
-                    'filename' => 'SomeRandomFile.txt',
-                    'description' => 'Simple description',
-                    'content_type' => 'text/plain',
-                ],
-            ],
-            'custom_fields' => [
-                ['id' => 25, 'value' => 'Second Custom Field'],
-            ],
-        ];
-
-        // Create the used mock objects
-        $client = $this->createMock(Client::class);
-        $client->expects($this->once())
-            ->method('requestPost')
-            ->with(
-                '/issues.xml',
-                $this->logicalAnd(
-                    $this->stringStartsWith('<?xml version="1.0"?>' . "\n" . '<issue>'),
-                    $this->stringEndsWith('</issue>' . "\n"),
-                    $this->stringContains('<watcher_user_ids type="array">'),
-                    $this->stringContains('</watcher_user_ids>'),
-                    $this->stringContains('<watcher_user_id>5</watcher_user_id>'),
-                    $this->stringContains(
-                        '<upload>'
-                        . '<token>first-token</token>'
-                        . '<filename>SomeRandomFile.txt</filename>'
-                        . '<description>Simple description</description>'
-                        . '<content_type>text/plain</content_type>'
-                        . '</upload>'
-                    ),
-                    $this->stringContains(
-                        '<custom_field id="25">'
-                        . '<value>Second Custom Field</value>'
-                        . '</custom_field>'
-                    ),
-                    $this->stringContains('<subject>Issue subject with some xml entities: &amp; &lt; &gt; " \' </subject>'),
-                    $this->stringContains('<description>Description with some xml entities: &amp; &lt; &gt; " \' </description>')
-                )
-            );
-
-        // Create the object under test
-        $api = new Issue($client);
-
-        // Perform the tests
-        $api->create($parameters);
     }
 
     /**
